+++ site/SectionHeading.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
-import { TocHeading } from "site/formatting"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faArrowDown } from "@fortawesome/free-solid-svg-icons/faArrowDown"
+import { TocHeading } from "clientUtils/owidTypes"
 
 export const SectionHeading = ({
     title,
